Requirements and setup updates

+++ opennft/opennft.py
@@ -844,7 +844,7 @@ class OpenNFT(QWidget):
         # Main logic
         # data preprocessing
         if config.USE_YIELD:
-            self.call_timer.setInterval(config.MAIN_LOOP_CALL_PERIOD / 3)
+            self.call_timer.setInterval(np.int32(config.MAIN_LOOP_CALL_PERIOD / 3))
             prepr_vol_state = self.eng.preprVol(fname, self.iteration, background=True, nargout=0)
             while not prepr_vol_state.done():
                 yield
